+++ app/src/main/java/de/triology/universeadm/user/imports/CSVParserImpl.java
@@ -4,8 +4,6 @@ import com.opencsv.CSVReader;
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
 import com.opencsv.exceptions.*;
-import org.apache.commons.io.input.TeeInputStream;
-import org.apache.commons.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,8 +11,6 @@ import javax.validation.constraints.NotNull;
 import java.io.*;
 import java.lang.reflect.Field;
 import java.util.*;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -24,26 +20,23 @@ import java.util.stream.Stream;
 public class CSVParserImpl implements CSVParser {
 
     private CsvToBean<CSVUserDTO> csv2bean;
-    private final AtomicReference<String[]> header = new AtomicReference<>();
+    private String[] header = null;
     private static final Logger logger =
             LoggerFactory.getLogger(CSVParserImpl.class);
 
 
-    public Stream<CSVUserDTO> parse(@NotNull InputStream fileStream) throws CsvRequiredFieldEmptyException, IOException {
+    public Stream<CSVUserDTO> parse(@NotNull InputStream fileStream) throws CsvRequiredFieldEmptyException {
 
-        PipedInputStream headerIn = new PipedInputStream();
-        TeeInputStream contentIn = new TeeInputStream(fileStream, new PipedOutputStream(headerIn));
+        BufferedReader contentReader = new BufferedReader(new InputStreamReader(fileStream));
+        CSVReader csvReader = new CsvLineNumberReader(contentReader, (String[] h) -> this.header = h);
 
-        BufferedReader contentReader = new BufferedReader(new InputStreamReader(contentIn));
-
-        CsvToBean<CSVUserDTO> beans = new CsvToBeanBuilder<CSVUserDTO>(new CsvLineNumberReader(contentReader))
+        CsvToBean<CSVUserDTO> beans = new CsvToBeanBuilder<CSVUserDTO>(csvReader)
                 .withType(CSVUserDTO.class)
                 .withIgnoreEmptyLine(true)
                 .withThrowExceptions(false)
                 .build();
 
         this.csv2bean = beans;
-        this.setHeader(headerIn);
 
         logger.debug("Build CsvToBean instance for CSVUserDTO");
 
@@ -90,7 +83,7 @@ public class CSVParserImpl implements CSVParser {
                         .withLineNumber(e.getLineNumber())
                         .withErrorMessage(e.getMessage());
 
-        Optional<String[]> optHeader = Optional.ofNullable(this.header.get());
+        Optional<String[]> optHeader = Optional.ofNullable(this.header);
 
         if (!optHeader.isPresent() || (optHeader.get().length != e.getLine().length)) {
             return createImportBuilder
@@ -134,25 +127,4 @@ public class CSVParserImpl implements CSVParser {
                 .apply(ImportError.Code.PARSING_ERROR)
                 .build();
     }
-
-    private void setHeader(InputStream in) {
-        Executors.newCachedThreadPool().execute(() -> {
-            try {
-                BufferedReader headerReader = new BufferedReader(new InputStreamReader(in));
-                CSVReader csvReader = new CSVReader(headerReader);
-                String[] header = csvReader.readNext();
-                // Append LineNumber column to header of CSV file
-                String[] headerWithLineColumn = ArrayUtils.add(header, CsvLineNumberReader.LINE_COLUMN);
-
-                headerReader.close();
-
-                this.header.set(headerWithLineColumn);
-
-                logger.debug("Read header with columns {}", Arrays.toString(this.header.get()));
-            } catch (CsvValidationException | IOException e) {
-                throw new RuntimeException(e);
-            }
-        });
-    }
-
 }